Subject fix: repair autograd correctness, matrix integrity, and softmax forward/backward
Body
Matrix
- fix copy_from: all three overloads now set this->size = data.size();
previously size stayed 0 after copy, breaking get_size() checks everywhere
- fix Matrix(vector, shape): run verifyShape on the arguments before assigning,
not after (data/shape were empty at check time)
- replace all assert("string" && cond) with throw runtime_error(!cond);
string-literal && always evaluates true, conditions were never enforced
- fix computeBroadcastResultShape: guard was inverted (!assertBroadcast → throw)
- fix transpose(shape_t perm): condition was perm.size()==ndims (passes when equal);
corrected to throw when perm.size() != ndims
- replace assert(1 && ...) in transpose_1D with throw (always passed, never fired)
- fix Matrix(const Matrix* two) and copy_from(ptr): error was constructed but not thrown
- fix matProduct2D: M/K/N now read from last two shape dims, not shape[0/1];
add lhsStart/rhsStart/resStart offsets to BLAS calls and generic fallback
- fix matmul recursive: pass rhsStart computed from rhs.numElementsSeen;
previously all batch slices multiplied from offset 0 of rhs
- fix transpose internal: replace index-reversal algorithm with explicit perm mapping;
reversal only correct for default transpose, produced OOB reads on other perms

Autograd / Tensor
- fix zero_grad: backOp nulled before op->backward in Tensor::backward;
moved null to after the backward call to avoid severing graph mid-pass
- fix tensor_overloads: remove right->frontOp = op from all binary operators;
setting frontOp on both operands clobbered weight/bias frontOp each forward pass
- fix Tensor copy constructor: no longer copies backOp/frontOp from source;
copying graph pointers into new tensors caused aliased backward paths
- rewrite cross_entropy and binary_cross_entropy using graph ops only;
previous version used ->at({i}) indexing which broke gradient flow

New operations
- add SoftmaxOperation: fused numerically stable row-wise softmax;
forward subtracts per-row max before exp, backward uses analytic
Jacobian-vector product s*(G - dot(G,s)) — no full Jacobian materialized
- add SumAxisOperation: graph-aware reduction along a given axis;
backward broadcasts upstream grad back to input shape by repeating
values along the summed axis
- replace Tensor::softmax() primitive decomposition with SoftmaxOperation;
old version called sum() over all elements, rows summed to ~0.1 not 1.0
- replace Tensor::sum(axis) off-graph make_tensor with SumAxisOperation

Optimizer / Module
- fix Adam/AdamW: m/v initialised only for non-zero-grad params but indexed
by full parameter index; now always push for all params to keep indices aligned
- fix Module::submodules: changed vector type and register_module signature
to use raw pointer; fixes object slicing in NeuralLayer forward loop
- fix NeuralLayer forward: was iterating by value (Module mod) causing slicing;
changed to iterate by pointer (auto mod)
- add gradient clipping to Optimizer::step() before update

Overload
- replace assert(size==) with throw in all vector binary ops
- fix division by zero guards: replace find(...,(double)0) with per-element
check using T(0) to avoid type mismatch and O(n) pre-scan overhead

Validation
- verified on MNIST (10-sample overfit): loss 25% → 2% CE in 1000 iters;
softmax rows now sum to 1.0, gradient flow confirmed end-to-end

+++ include/DataLoader/Dataset.hpp
@@ -40,8 +40,8 @@ public:
         if (idx >= size())
             throw std::out_of_range("Dataset index out of range.");
 
-        long n_feat = X.shape[1];
-        long n_out  = Y.shape[1];
+        size_t n_feat = X.shape[1];
+        size_t n_out  = Y.shape[1];
 
         std::vector<T> x_row(X.data.begin() + idx * n_feat,
                               X.data.begin() + idx * n_feat + n_feat);
@@ -108,10 +108,10 @@ public:
         // Returns {X_batch [batch, n_feat], Y_batch [batch, n_out]}
         std::pair<Matrix<T>, Matrix<T>> operator*() const {
             size_t end  = std::min(cursor + loader.batch_size, loader.dataset.size());
-            size_t n    = end - cursor;
+            size_t n  = end - cursor;
 
-            long n_feat = loader.dataset.X.shape[1];
-            long n_out  = loader.dataset.Y.shape[1];
+            size_t n_feat = loader.dataset.X.shape[1];
+            size_t n_out  = loader.dataset.Y.shape[1];
 
             std::vector<T> xb, yb;
             xb.reserve(n * n_feat);
@@ -125,8 +125,8 @@ public:
             }
 
             return {
-                Matrix<T>(xb, {(long)n, n_feat}),
-                Matrix<T>(yb, {(long)n, n_out})
+                Matrix<T>(xb, {n, n_feat}),
+                Matrix<T>(yb, {n, n_out})
             };
         }
     };
